apps/api/pyproject.toml
[project]
name = "cipherchat-api"
version = "0.1.0"
description = "FastAPI backend for CipherChat"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "email-validator>=2.2.0",
  "fastapi>=0.115.8",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.5",
  "pydantic-settings>=2.7.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.38",
  "uvicorn[standard]>=0.34.0"
]

[dependency-groups]
dev = [
  "black>=25.1.0",
  "httpx>=0.28.1",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.9.7"
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

'target-version' = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
